% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial Distribution}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of iterations}

\item{n}{Size of sample}

\item{p}{Probability of success}
}
\value{
Table of probabilities for each number of successes
}
\description{
Binomial Distribution
}
\examples{
\dontrun{mybin(iter=1000,n=18, p=0.3)}
}
