% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Sampling}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of iterations}

\item{x}{Sample size}

\item{fun}{What data to sample}

\item{alpha}{Quantiles to find}

\item{cx}{Font size}

\item{...}{Other parameters}
}
\value{
List of values
}
\description{
Bootstrap Sampling
}
\examples{
  \dontrun{myboot2(x=sam,fun="sd",alpha=0.1,xlab="sd")}
}
