% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Confidence Interval}
\usage{
myci(x, ci = 0.95)
}
\arguments{
\item{x}{A vector of data}

\item{ci}{Confidence level of the interval}
}
\value{
Confidence interval of the data
}
\description{
Confidence Interval
}
\examples{
  \dontrun{myci(x=c(5.0581, 4.9707, 5.0893, 4.9334, 4.9777, 5.0285, 4.8555, 4.9565))}
}
