% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{Uniform Distribution}
\usage{
myclt(n, iter)
}
\arguments{
\item{n}{Sample size}

\item{iter}{Number of iteration}
}
\value{
A histogram of the data and a vector containing the sum of the data
}
\description{
Uniform Distribution
}
\examples{
\dontrun{w=myclt(n=10,iter=10000)}
}
