% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycurve.R
\name{mycurve}
\alias{mycurve}
\title{Normal Distribution}
\usage{
mycurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of values}

\item{sigma}{Standard deviation of values}

\item{a}{Range of values from (-infinity, a)}
}
\value{
A plot with the area to a and a list containing the mean, standard deviation,
 and area
}
\description{
Normal Distribution
}
\examples{
\dontrun{mycurve(10,5,3)}
}
