% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Max Likelihood}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{Function to apply to x and param}

\item{x}{Vector containing data}

\item{param}{Parameters for the plot}

\item{...}{Other parameters}
}
\value{
A plot with a point at the max likelihood
}
\description{
Max Likelihood
}
\examples{
\dontrun{mymaxlik(x=c(9,9,1,9,9,9),param=seq(0,1,length=1000),lfun=logbin,
  xlab=expression(pi),main="Binomial",cex.main=2)}
}
