% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Optimal Number of Tickets Sold}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats}

\item{gamma}{Probability of overbooking}

\item{p}{Probability of a "show"}
}
\value{
A list of values containing values of discrete distribution and normal
approximation, p, and gamma
}
\description{
Optimal Number of Tickets Sold
}
\examples{
  \dontrun{ntickets(N=400,gamma = 0.02, p = 0.95)}
}
